requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions
face_recognition
pillow